move Lazy type to helper_types
